+++ CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [2.1.0] - 2025-06-30
+
+### Added
+- 4 new functions for converse status management ```getExternalConversationHistory``` ```getExternalConversationHistory``` ```clearExternalConversationHistory``` ```saveExternalConversationHistory```
+- New optional props ```onNewTextMessage``` for handling all text messages coming in while the component is running.
+- New optional property ```initSaveExternalConversation``` in object ```ZowieConfig``` to support automation ```saveExternalConversationHistory```
+
 ## [2.0.4] - 2025-06-10
 
 ### Fixed